requirements.txt
# =============================================================================
# VERIHOME - DEPENDENCIAS DE PYTHON
# =============================================================================

# Framework principal
Django==4.2.7
django-extensions==3.2.3

# Autenticación y autorización
django-allauth==0.57.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Base de datos
psycopg2-binary==2.9.9
django-migration-testcase==0.0.15

# Cache y tareas asíncronas
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
flower==2.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.0

# API y CORS
django-cors-headers==4.3.1
django-filter==23.5

# Configuración de entorno
python-decouple==3.8
python-dotenv==1.0.0

# Archivos y media
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34

# Servidor web
whitenoise==6.6.0
gunicorn==21.2.0

# Monitoreo y logging
sentry-sdk==1.40.0
django-debug-toolbar==4.2.0

# Pagos
stripe==7.8.0

# Utilidades
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
setuptools>=80.0.0

# Desarrollo y testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Documentación API
drf-spectacular==0.26.5

# Seguridad
cryptography==41.0.7

# WebSockets y Real-time
channels==4.2.2
channels-redis==4.2.0
daphne==4.1.2

# Procesamiento de datos
pandas==2.1.4
numpy==1.25.2

# Mapas
geopy==2.4.1

# Validación de datos
cerberus==1.3.5

# Archivos CSV y Excel
openpyxl==3.1.2

# Generación de PDFs
reportlab==4.0.7
weasyprint==60.2 